+++ lib/src/helpers/settings/month_picker_settings.dart
@@ -5,7 +5,8 @@ class MonthPickerDialogSettings {
   const MonthPickerDialogSettings({
     this.dialogSettings = defaultPickerdialogSettings,
     this.headerSettings = defaultPickerHeaderSettings,
-    this.buttonsSettings = defaultPickerbuttonsSettings,
+    this.dateButtonsSettings = defaultPickerDateButtonsSettings,
+    this.actionBarSettings = defaultPickerActionBarSettings,
   });
 
   ///The customizations of the dialog part of the package.
@@ -14,8 +15,11 @@ class MonthPickerDialogSettings {
   ///The customizations of the header part of the package.
   final PickerHeaderSettings headerSettings;
 
-  ///The customizations of the buttons part of the package.
-  final PickerButtonsSettings buttonsSettings;
+  ///The customizations of the date buttons part of the package.
+  final PickerDateButtonsSettings dateButtonsSettings;
+
+  ///The customizations of the date buttons part of the package.
+  final PickerActionBarSettings actionBarSettings;
 }
 
 ///The default settings for the month picker.