+++ lib/src/pickers/show_month_picker.dart
@@ -19,12 +19,6 @@ import '/month_picker_dialog.dart';
 ///
 /// `yearStylePredicate:` allows you to individually customize each year.
 ///
-/// `confirmWidget:` lets you set a custom confirm widget.
-///
-/// `cancelWidget:` lets you set a custom cancel widget.
-///
-/// `customDivider:` lets you add a custom divider between the months/years and the confirm/cancel buttons.
-///
 /// `headerTitle:` lets you add a custom title to the header of the dialog (default is `null`).
 ///
 /// `monthPickerDialogSettings:` is the object that will hold all of the style of the picker dialog (default is `defaultMonthPickerDialogSettings`).
@@ -40,9 +34,6 @@ Future<DateTime?> showMonthPicker({
   bool Function(int)? selectableYearPredicate,
   ButtonStyle? Function(DateTime)? monthStylePredicate,
   ButtonStyle? Function(int)? yearStylePredicate,
-  Widget? confirmWidget,
-  Widget? cancelWidget,
-  Widget? customDivider,
   Widget? headerTitle,
   MonthPickerDialogSettings monthPickerDialogSettings =
       defaultMonthPickerDialogSettings,
@@ -58,11 +49,8 @@ Future<DateTime?> showMonthPicker({
     selectableYearPredicate: selectableYearPredicate,
     monthStylePredicate: monthStylePredicate,
     yearStylePredicate: yearStylePredicate,
-    confirmWidget: confirmWidget,
-    cancelWidget: cancelWidget,
     theme: theme,
     useMaterial3: theme.useMaterial3,
-    customDivider: customDivider,
     headerTitle: headerTitle,
     onlyYear: onlyYear,
   );